chore: move essential methods from screen/Dashboard.js to App.js
